(refactor): move sorting logic to usecase
